+++ WindowsFormsApp9/Form1.cs
@@ -1,15 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.ComponentModel;
-using System.Data;
 using System.Diagnostics;
 using System.Drawing;
-using System.Linq;
-using System.Text;
 using System.Threading;
 using System.Windows.Forms;
-using System.Linq;
-using System.Runtime.InteropServices;
 
 namespace WindowsFormsApp9
 {
@@ -19,7 +13,7 @@ namespace WindowsFormsApp9
         {
             InitializeComponent();
             Taskbar.Hide();
-            FormBorderStyle = System.Windows.Forms.FormBorderStyle.None;
+            FormBorderStyle = FormBorderStyle.None;
             WindowState = FormWindowState.Normal;
             StartPosition = FormStartPosition.Manual;
             Location = new Point(0, 0);
@@ -55,69 +49,10 @@ namespace WindowsFormsApp9
 
         }
 
-        public class Taskbar
-        {
-            [DllImport("user32.dll")]
-            private static extern int FindWindow(string className, string windowText);
-
-            [DllImport("user32.dll")]
-            private static extern int ShowWindow(int hwnd, int command);
-
-            [DllImport("user32.dll")]
-            public static extern int FindWindowEx(int parentHandle, int childAfter, string className, int windowTitle);
-
-            [DllImport("user32.dll")]
-            private static extern int GetDesktopWindow();
-
-            private const int SW_HIDE = 0;
-            private const int SW_SHOW = 1;
-
-            protected static int Handle
-            {
-                get
-                {
-                    return FindWindow("Shell_TrayWnd", "");
-                }
-            }
-
-            protected static int HandleOfStartButton
-            {
-                get
-                {
-                    int handleOfDesktop = GetDesktopWindow();
-                    int handleOfStartButton = FindWindowEx(handleOfDesktop, 0, "button", 0);
-                    return handleOfStartButton;
-                }
-            }
-
-            private Taskbar()
-            {
-                // hide ctor
-            }
-
-            public static void Show()
-            {
-                ShowWindow(Handle, SW_SHOW);
-                ShowWindow(HandleOfStartButton, SW_SHOW);
-            }
-
-            public static void Hide()
-            {
-                ShowWindow(Handle, SW_HIDE);
-                ShowWindow(HandleOfStartButton, SW_HIDE);
-            }
-        }
-
         public void WorkThreadFunction(Screen screen)
         {
             try
             {
-                if (screen.Primary == true)
-                {
-                   
-                    
-                }
-
                 if (screen.Primary == false)
                 {
                     int mostLeft = screen.WorkingArea.Left;
@@ -128,7 +63,7 @@ namespace WindowsFormsApp9
                         form.WindowState = FormWindowState.Normal;
                         form.StartPosition = FormStartPosition.Manual;
                         form.Location = new Point(mostLeft, mostTop);
-                        form.FormBorderStyle = System.Windows.Forms.FormBorderStyle.None;
+                        form.FormBorderStyle = FormBorderStyle.None;
                         form.Size = new Size(screen.Bounds.Width, screen.Bounds.Height);
                         form.BackColor = Color.Black;
                         form.ShowDialog();
@@ -139,23 +74,6 @@ namespace WindowsFormsApp9
             {
                 // log errors
             }
-        }
-
-
-        private void button1_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void pictureBox1_Click(object sender, EventArgs e)
-        {
-
-        }
-
-
-        private void label2_Click(object sender, EventArgs e)
-        {
-
         }
 
         protected override CreateParams CreateParams
@@ -172,12 +90,12 @@ namespace WindowsFormsApp9
         private void pictureBox2_Click(object sender, EventArgs e)
         {
             Taskbar.Show();
-            System.Windows.Forms.Application.Exit();
+            Application.Exit();
         }
 
         protected override void OnClosing(CancelEventArgs e)
         {
-                Taskbar.Show();
+            Taskbar.Show();
             base.OnClosing(e);
         }
 
@@ -186,23 +104,18 @@ namespace WindowsFormsApp9
             Console.WriteLine(textBox2);
         }
 
-        private void label1_Click(object sender, EventArgs e)
-        {
-
-        }
-
         private void pictureBox3_Click(object sender, EventArgs e)
         {
             Taskbar.Show();
-            System.Windows.Forms.Application.Exit();
+            Application.Exit();
         }
 
         private void button1_Click_1(object sender, EventArgs e)
         {
             Taskbar.Show();
-            System.Windows.Forms.Application.Exit();
+            Application.Exit();
         }
     }
-    
+
 
 }
